requirements.txt
# Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0

# Config & validation
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data / ML
numpy>=1.26.0
openai>=1.40.0
mlflow>=2.14.0

# Storage (shared KB for multi-replica microservices)
redis>=5.0.0

# Tests
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
fakeredis>=2.20.0
respx>=0.21.0
